backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
motor>=3.3.2
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.1
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
email-validator>=2.1.0>=0.109.0